requirements.txt
numpy==1.26.4
opencv-contrib-python==4.10.0.84
flask==3.0.3
Flask-PyMongo==2.3.0
celery==5.4.0
python-dotenv==1.0.1
redis==5.2.0